updated version of dapper

+++ Dapper.java
@@ -1,10 +1,10 @@
 public class Dapper
 {
-    public float density(GraphNetwork network){
+    public double density(GraphNetwork network){
         int numNodes = network.countNode(network);
         int numEdges = network.countEdges(network);
 
-        float density = numEdges / (numNodes*(numNodes -1));
+        double density = (double) numEdges / (numNodes*(numNodes -1));
         
         return density;
     }
